Updates to Delete Call
