+++ OSFOLCrossPlatform/OSFOLCrossPlatform/Infrastructure/NonNullToBooleanConverter.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Globalization;
+using Xamarin.Forms;
+
+namespace OSFOLCrossPlatform.Infrastructure
+{
+    public class NonNullToBooleanConverter : IValueConverter
+    {
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            if (value is string)
+            {
+                return !string.IsNullOrEmpty((string)value);
+            }
+
+            return value != null;
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            return null;
+        }
+    }
+}
