+++ OSFOLCrossPlatform/OSFOLCrossPlatform/OSFOLCrossPlatform.csproj
@@ -36,8 +36,10 @@
     <Compile Include="Controls\BindablePicker.cs" />
     <Compile Include="Data\ExpenseDatabase.cs" />
     <Compile Include="ExpensePhotoPreview.cs" />
+    <Compile Include="Infrastructure\NonNullToBooleanConverter.cs" />
     <Compile Include="Infrastructure\ObservableObject.cs" />
     <Compile Include="Model\IImage.cs" />
+    <Compile Include="Pages\ExpenseSetsCarouselPage.cs" />
     <Compile Include="ViewModels\AddExpenseViewModel.cs" />
     <Compile Include="Model\BuildExpense.cs" />
     <Compile Include="Model\ExpenseModel.cs" />
@@ -47,7 +49,6 @@
     <Compile Include="Views\AddExpenseView.xaml.cs">
       <DependentUpon>AddExpenseView.xaml</DependentUpon>
     </Compile>
-    <Compile Include="AddExpenseCS.cs" />
     <Compile Include="App.cs" />
     <Compile Include="Views\AddquickExpenseView.xaml.cs">
       <DependentUpon>AddquickExpenseView.xaml</DependentUpon>
@@ -55,7 +56,7 @@
     <Compile Include="Views\EditExpense.xaml.cs">
       <DependentUpon>EditExpense.xaml</DependentUpon>
     </Compile>
-    <Compile Include="EditExpenseCS.cs" />
+    <Compile Include="Pages\EditExpenseCS.cs" />
     <Compile Include="Views\ExpensesViewCell.cs" />
     <Compile Include="Views\LoginPage.xaml.cs">
       <DependentUpon>LoginPage.xaml</DependentUpon>
@@ -64,10 +65,7 @@
       <DependentUpon>MainPage.xaml</DependentUpon>
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Views\Report.xaml.cs">
-      <DependentUpon>Report.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="ExpensesPage.cs" />
+    <Compile Include="Pages\ExpensesPage.cs" />
     <Compile Include="Model\Login.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -127,12 +125,6 @@
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Views\EditExpense.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </EmbeddedResource>
-  </ItemGroup>
-  <ItemGroup>
-    <EmbeddedResource Include="Views\Report.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </EmbeddedResource>
