+++ OSFOLCrossPlatform/OSFOLCrossPlatform/Pages/ExpenseSetsCarouselPage.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection.Emit;
+using System.Text;
+
+using Xamarin.Forms;
+
+namespace OSFOLCrossPlatform.Pages
+{
+    public class ExpenseSetsCarouselPage : CarouselPage
+    {
+        public ExpenseSetsCarouselPage()
+        {
+            //Title = "Expense Set {Binding ExpenseSet}";
+            Title = "Expenses";
+
+
+        }
+    }
+}
