+++ OSFOLCrossPlatform/OSFOLCrossPlatform/Views/ExpensesViewCell.cs
@@ -3,8 +3,10 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection.Emit;
 using System.Text;
+using OSFOLCrossPlatform.Model;
 
 using Xamarin.Forms;
+using System.Threading.Tasks;
 
 namespace OSFOLCrossPlatform.Views
 {
@@ -20,31 +22,31 @@ namespace OSFOLCrossPlatform.Views
             #endregion
 
             #region Create Expense Stack
-            //var expenseLabel = new Label
-            //{
-            //    Text = "Expense",
-            //    FontAttributes = FontAttributes.Bold
-            //};
-            //var expense = new Label();
-            //expense.SetBinding(Label.TextProperty, "Expense");
+            var expenseLabel = new Label
+            {
+                Text = "Expense",
+                FontAttributes = FontAttributes.Bold
+            };
+            var expense = new Label();
+            expense.SetBinding(Label.TextProperty, "ExpenseDetails");
 
-            //var texpenseStack = new StackLayout
-            //{
-            //    Children = {
-            //        expenseLabel,
-            //        expense
-            //    }
-            //};
+            var expenseStack = new StackLayout
+            {
+                Children = {
+                    expenseLabel,
+                    expense
+                }
+            };
             #endregion
 
             #region Create Customer Stack
             var customerLabel = new Label
             {
-                Text = "Customers",
+                Text = "CustomerID",
                 FontAttributes = FontAttributes.Bold
             };
             var customer = new Label();
-            customer.SetBinding(Label.TextProperty, "Customers");
+            customer.SetBinding(Label.TextProperty, "CustomersID");
 
             var customerStack = new StackLayout
             {
@@ -58,11 +60,11 @@ namespace OSFOLCrossPlatform.Views
             #region Create Opportunity Stack
             var opportunityLabel = new Label
             {
-                Text = "Opportunity",
+                Text = "SaleOpportunityID",
                 FontAttributes = FontAttributes.Bold
             };
             var opportunity = new Label();
-            opportunity.SetBinding(Label.TextProperty, "Opportunity");
+            opportunity.SetBinding(Label.TextProperty, "SaleOpportunityID");
 
             var opportunityStack = new StackLayout
             {
@@ -71,6 +73,29 @@ namespace OSFOLCrossPlatform.Views
                     opportunity
                 }
             };
+            #endregion
+
+            #region Create MenuItem
+            var deleteAction = new MenuItem
+            {
+                Text = "Delete",
+                IsDestructive = true
+            };
+            deleteAction.SetBinding(MenuItem.CommandProperty, "DeleteActionSelected");
+
+            deleteAction.Clicked += async (sender, e) =>
+            {
+                var menuItem = (MenuItem)sender;
+                Expense thisModel = ((Expense)menuItem.BindingContext);
+                App.Database.DeleteItem(thisModel.ExpenseID);
+
+                //Wait for the iOS animation to finish
+                if (Device.OS == TargetPlatform.iOS)
+                    await Task.Delay(300);
+
+                MessagingCenter.Send<object>(this, "RefreshData");
+            };
+            ContextActions.Add(deleteAction);
             #endregion
 
             StackLayout cellStack;
@@ -86,7 +111,7 @@ namespace OSFOLCrossPlatform.Views
                     Orientation = StackOrientation.Horizontal,
                     Children = {
                         lighthouseImage,
-                        customerStack
+                        expenseStack
                     }
                 };
             }
@@ -103,6 +128,7 @@ namespace OSFOLCrossPlatform.Views
                     Orientation = StackOrientation.Horizontal,
                     Children = {
                         lighthouseImage,
+                        expenseStack,
                         customerStack,
                         opportunityStack
                     }
