+++ OSFOLCrossPlatform/OSFOLCrossPlatform/Views/MasterNavPage.xaml
@@ -1,0 +1,16 @@
+﻿<MasterNavPage xmlns="http://xamarin.com/schemas/2014/forms"
+                  xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+                  xmlns:local="clr-namespace:OSFOLCrossPlatform;assembly=OSFOLCrossPlatform"
+                  x:Class="OSFOLCrossPlatform.Views.MasterNavPage"
+                  MasterBehaviour="Split">
+  <MasterNavPage.Master>
+    <local:MasterPage x:Name="masterPage" />
+  </MasterNavPage.Master>
+  <MasterNavPage.Detail>
+    <NavigationPage>
+      <x:Arguments>
+        <local:ContactsPage />
+      </x:Arguments>
+    </NavigationPage>
+  </MasterNavPage.Detail>
+</MasterNavPage>
