+++ OSFOLCrossPlatform/OSFOLCrossPlatform/Views/MasterNavPage.xaml.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+using Xamarin.Forms;
+
+namespace OSFOLCrossPlatform.Views
+{
+    public partial class MasterNavPage : ContentPage
+    {
+        public MasterNavPage()
+        {
+            InitializeComponent();
+        }
+    }
+}
